+++ modules/demux/mp4/libmp4.h
@@ -768,8 +768,8 @@ typedef struct
     uint16_t i_samplesize;
     uint16_t i_compressionid;
     uint16_t i_reserved3;
-    uint32_t i_sampleratehi; /* timescale of track */
-    uint32_t i_sampleratelo;
+    uint16_t i_sampleratehi; /* timescale of track */
+    uint16_t i_sampleratelo;
 
     /* for version 1 (i_reserved1[0] == 1) */
     uint32_t i_sample_per_packet;
